+++ CHANGELOG.md
@@ -37,6 +37,7 @@ _Not yet on NuGet..._
 * WinForms: Exposed `SKControl` so users may bind to its events (#3819) @CD-SailingPerf
 * Scatter: Added support for `Scale` and `Offset` properties (#3835) @bukkideme
 * Axis Lines: Separated `LegendText` from `LabelText` so items may be configured separately
+* Heatmap: Exposed `CellWidth` and `CellHeight` as an alternative sizing strategy to setting `Extent` (#3869) @alexisvrignaud
 
 ## ScottPlot 5.0.34
 _Published on [NuGet](https://www.nuget.org/profiles/ScottPlot) on 2024-05-05_